Created the chart and it showcases linear growth

+++ ComplexityExplorer/DataPoint.swift
@@ -10,6 +10,6 @@ import Foundation
 struct DataPoint: Identifiable {
     let id = UUID()
     var x: Int
-    var y: Int
+    var y: Double
     let rate: String
 }
